% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{finnsq}
\alias{finnsq}
\title{Takes in a numeric value and squares it.}
\usage{
finnsq(x)
}
\arguments{
\item{x}{A numeric value to be squared}
}
\value{
The square of the input
}
\description{
Takes in a numeric value and squares it.
}

