% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonnotestest.R
\name{getpnotes}
\alias{getpnotes}
\title{Gets notes from lootbox api and transfers them into a data frame format}
\usage{
getpnotes()
}
\arguments{
\item{no}{paramater needed}
}
\value{
a datafram containig the json information
}
\description{
Gets notes from lootbox api and transfers them into a data frame format
}

