% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonnotestest.R
\name{refreq}
\alias{refreq}
\title{Request from refence server}
\usage{
refreq(url = "http://0.0.0.0:5000/", end = "units/search", req = { })
}
\arguments{
\item{url}{= the url you want the data to come from i.e. "http://example.com/api/"}

\item{end}{= the prefex to the above url i.e. "example/"}

\item{req}{= the request you want to send}
}
\value{
a list containig the information
}
\description{
Sends a request to the framespace reference server and retreaves the content of the response
}

